backend/pyproject.toml
[tool.poetry]
name = "PRAI"
version = "0.1.0"
description = ""
authors = ["Max Takacs <max@mraisolutions.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.12"  # Matching your Dockerfile's Python version
fastapi = "*"
uvicorn = "*"
python-dotenv = "*"
openai = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"